Function validator_push_cdn::infra::libp2p_network_from_config

async fn libp2p_network_from_config<TYPES: NodeType>(
    config: NetworkConfig<TYPES::SignatureKey, TYPES::ElectionConfigType>,
    pub_key: TYPES::SignatureKey
) -> Libp2pNetwork<Message<TYPES>, TYPES::SignatureKey>

Create a libp2p network from a config file and public key

Panics

If unable to create bootstrap nodes multiaddres or the libp2p config is invalid