Function validator_push_cdn::infra::load_config_from_file

pub fn load_config_from_file<TYPES: NodeType>(
    config_file: &str
) -> NetworkConfig<TYPES::SignatureKey, TYPES::ElectionConfigType>

Reads a network configuration from a given filepath

Panics

if unable to convert the config file into toml

Note

This derived config is used for initialization of orchestrator, therefore known_nodes_with_stake will be an initialized vector full of the node’s own config. my_own_validator_config will be generated from seed here for loading config from orchestrator, or else it will be loaded from file.